+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## HEAD
 
 - Switch dependency from `fetch-super-headers` to `@mjackson/headers`
+- Use `for await...of` to iterate over `ReadableStream` internally. This will also cancel the stream when the loop exits from e.g. an error in a user-defined `part` handler.